refactor: moved db dependent commands to cmds package
