+++ cmds/update.go
@@ -1,12 +1,13 @@
-package db
+package cmds
 
 import (
 	"errors"
 
+	db "github.com/joshmedeski/sesh/database"
 	"github.com/urfave/cli/v2"
 )
 
-func (c *SqliteDatabase) Update() *cli.Command {
+func Update(storage db.Storage) *cli.Command {
 	return &cli.Command{
 		Name:                   "update",
 		Aliases:                []string{"u"},
@@ -18,7 +19,7 @@ func (c *SqliteDatabase) Update() *cli.Command {
 			if len(args) != 3 {
 				return errors.New("Invalida args") // TODO improve error message
 			}
-			return c.UpdateEntry(args[0], args[1], args[2])
+			return storage.UpdateEntry(args[0], args[1], args[2])
 		},
 	}
 }
