+++ seshcli/seshcli.go
@@ -1,26 +1,21 @@
 package seshcli
 
 import (
-	"os"
-
 	"github.com/joshmedeski/sesh/cmds"
-	db "github.com/joshmedeski/sesh/database"
 
+	db "github.com/joshmedeski/sesh/database"
 	"github.com/urfave/cli/v2"
 )
 
-func App(version string) cli.App {
-	sqlPath := os.ExpandEnv("$HOME/.local/share/sesh/sesh.db")
-	storage := db.NewSqliteDatabase(sqlPath)
-
+func App(version string, storage db.Storage) cli.App {
 	return cli.App{
 		Name:    "sesh",
 		Version: version,
 		Usage:   "Smart session manager for the terminal",
 		Commands: []*cli.Command{
-			storage.Add(),
-			storage.Delete(),
-			storage.Update(),
+			cmds.Add(storage),
+			cmds.Delete(storage),
+			cmds.Update(storage),
 			cmds.List(),
 			cmds.Choose(),
 			cmds.Connect(),
